moved logger.py to utils
